+++ source/sdl/mmus2mid.h
@@ -53,7 +53,7 @@ typedef struct MIDI                    /* a midi file */
    int divisions;                      /* number of ticks per quarter note */
    struct {
       unsigned char *data;             /* MIDI message stream */
-      int len;                         /* length of the track data */
+      size_t len;                      /* length of the track data */
    } track[MIDI_TRACKS]; 
 } MIDI;
 #endif // DJGPP
